Make some Font fields private
